Count modified binary files as one insertion and one deletion
Git diff --numstat gives no line information for modified binary files
which trips up the part of the prompt that aggregates modification
stats.

For these files I simply decided to add +1 to both the number of
insertions and to the number of deletions.  This makes sense if the
binary file is viewed as a single long line that is modified.